--- presentation/src/main/java/com/paykidscompose/presentation/model/ChatMessageUIModel.kt
@@ -1,6 +1,0 @@
-package com.paykidscompose.presentation.model
-
-data class ChatMessageUIModel(
-    val text: String,
-    val isFromGpt: Boolean
-): UIModel()
