+++ presentation/src/main/java/com/paykidscompose/presentation/model/study/ChatMessageUIModel.kt
@@ -1,0 +1,8 @@
+package com.paykidscompose.presentation.model.study
+
+import com.paykidscompose.presentation.model.UIModel
+
+data class ChatMessageUIModel(
+    val text: String,
+    val isFromGpt: Boolean
+): UIModel()
